Add files via upload
Based on Joel Wetzel iPhone WiFi Presence Sensor.  I just added the ability to turn it on and off manually (switch capability) as well as auto-turn off (this acts as an auto-update if they are still connected).  
Use this to control a Virtual Presence Sensor using my Virtual Switch Universal Device Type uDTH and select Presence Sensor in Preferences.